\name{MI_TSCA}

\docType{data}

\alias{MI_TSCA}

\title{Occurrence of Tsuga canadensis (Eastern hemlock) in Michigan}

\description{
 Eastern hemlock (\emph{Tsuga canadensis L.}) analyzed in
 Lany et al. (2019). The date comprise hemlock occurrence (binomial
 outcome) on 17,743 forest stands across Michigan, USA. A set of
 covariates were also observed at each stand and can be used to explain
 the probability of hemlock occurrence. Covariates included minimum
 winter temperature (MIN), maximum summer temperature (MAX), total
 precipitation in the coldest quarter of the year (WIP), total
 precipitation in the warmest quarter of the year (SUP), annual actual
 evapotranspiration (AET) and annual climatic water deficit
 (DEF). Spatial coordinates are recorded in 
  Longitude (long) and latitude (lat) which are Albers Equal Area (proj4string
"+proj=aea +lat_1=55 +lat_2=65 +lat_0=50 +lon_0=-154 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"). See
 Lany et al. (2019) for details. 
}

\usage{
data(MI_TSCA)
}

\references{
Lany, N. K., Zarnetske, P. L., Finley, A. O. and McCullough, D. G. 2019. Complimentary strengths of spatially-explicit and multi-species distribution models. – Ecography doi: 10.1111/ecog.04728

}

\format{
  A data frame containing 17,743 rows and 9 columns. 
}

\keyword{datasets}
