\name{print.spDiag}

\alias{print.spDiag}

\title{Methods for spDiag Object}

\description{
  Methods for extracting information from \code{spDiag}.
}

\usage{
\method{print}{spDiag}(x, \dots)
}

\arguments{
  \item{x}{object of class \code{spDiag}.}
  \item{\dots}{currently not used.}
}

\details{
  A standard extractor function for printing objects of class \code{spDiag}. 
}

\keyword{model}
